requirements.txt
absl-py==0.9.0
actionlib==1.12.0
angles==1.9.12
apturl==0.5.2
asn1crypto==0.24.0
astor==0.8.1
azure-iot-device==2.1.1
azure-iot-hub==2.2.0
azure-iothub-device-client==1.4.6
azure-iothub-service-client==1.4.6
bondpy==1.8.3
Brlapi==0.6.6
camera-calibration==1.14.0
camera-calibration-parsers==1.11.13
certifi==2018.1.18
chardet==3.0.4
command-not-found==0.3
controller-manager==0.17.0
controller-manager-msgs==0.17.0
cryptography==2.1.4
cupshelpers==1.0
cv-bridge==1.13.0
cycler==0.10.0
defer==1.0.6
diagnostic-analysis==1.9.3
diagnostic-common-diagnostics==1.9.3
diagnostic-updater==1.9.3
distro==1.5.0
distro-info===0.18ubuntu0.18.04.1
docutils==0.16
dynamic-reconfigure==1.6.3
gast==0.3.3
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.15
gennodejs==2.0.1
genpy==0.6.9
google-pasta==0.2.0
grpcio==1.28.1
h5py==2.10.0
httplib2==0.9.2
idna==2.6
image-geometry==1.13.0
imageai==2.1.5
interactive-markers==1.11.4
isodate==0.6.0
janus==0.4.0
joint-state-publisher==1.12.15
kdl-parser-py==1.13.1
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
keyring==10.6.0
keyrings.alt==3.0
kiwisolver==1.2.0
language-selector==0.1
laser-geometry==1.6.4
launchpadlib==1.10.6
lazr.restfulclient==0.13.5
lazr.uri==1.0.3
louis==3.5.0
macaroonbakery==1.1.3
Markdown==3.2.1
matplotlib==3.2.1
message-filters==1.14.5
msrest==0.6.13
netifaces==0.10.4
numpy==1.18.2
oauth==1.0.1
oauthlib==3.1.0
olefile==0.45.1
opencv-python==4.2.0.34
paho-mqtt==1.5.0
pexpect==4.2.1
Pillow==7.1.1
protobuf==3.11.3
pycairo==1.16.2
pycrypto==2.6.1
pycups==1.9.73
pygobject==3.26.1
pymacaroons==0.13.0
PyNaCl==1.1.2
pyparsing==2.4.7
pyRFC3339==1.0
PySocks==1.7.1
python-apt==1.6.5+ubuntu0.2
python-dateutil==2.8.1
python-debian==0.1.32
python-qt-binding==0.4.0
pytz==2018.3
pyxdg==0.25
PyYAML==5.3.1
qt-dotgraph==0.4.0
qt-gui==0.4.0
qt-gui-cpp==0.4.0
qt-gui-py-common==0.4.0
reportlab==3.4.0
requests==2.23.0
requests-oauthlib==1.3.0
requests-unixsocket==0.1.5
resource-retriever==1.12.6
scipy==1.4.1
SecretStorage==2.3.1
sensor-msgs==1.12.7
simplejson==3.13.2
six==1.14.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.3
system-service==0.3
systemd-python==234
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
tf==1.12.1
tf-conversions==1.12.1
tf2-geometry-msgs==0.6.5
tf2-kdl==0.6.5
tf2-py==0.6.5
tf2-ros==0.6.5
topic-tools==1.14.5
transitions==0.8.1
uamqp==1.2.6
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
urdfdom-py==0.4.2
urllib3==1.22
wadllib==1.3.2
Werkzeug==1.0.1
wrapt==1.12.1
xacro==1.13.5
xkit==0.0.0
zope.interface==4.3.2
